+++ src/main/java/org/truffleruby/interop/InteropNodes.java
@@ -58,6 +58,7 @@ import org.truffleruby.language.control.RaiseException;
 import org.truffleruby.shared.TruffleRuby;
 
 import java.io.IOException;
+import java.util.Arrays;
 
 @CoreClass("Truffle::Interop")
 public abstract class InteropNodes {
@@ -763,6 +764,23 @@ public abstract class InteropNodes {
         public Object toJavaArray(DynamicObject interopModule, DynamicObject array,
                                   @Cached("of(array)") ArrayStrategy strategy) {
             return getContext().getEnv().asGuestValue(strategy.newMirror(array).copyArrayAndMirror().getArray());
+        }
+
+        @Specialization(guards = "!isRubyArray(object)")
+        public Object coerce(DynamicObject interopModule, DynamicObject object) {
+            return null;
+        }
+
+    }
+
+    @Primitive(name = "to_java_list")
+    @ImportStatic(ArrayGuards.class)
+    public abstract static class InteropToJavaListNode extends PrimitiveArrayArgumentsNode {
+
+        @Specialization(guards = { "isRubyArray(array)", "strategy.matches(array)" }, limit = "STORAGE_STRATEGIES")
+        public Object toJavaList(DynamicObject interopModule, DynamicObject array,
+                                  @Cached("of(array)") ArrayStrategy strategy) {
+            return getContext().getEnv().asGuestValue(Arrays.asList(strategy.newMirror(array).getBoxedCopy()));
         }
 
         @Specialization(guards = "!isRubyArray(object)")